Ignore 409 error when reclaiming task when build is finished
We can ignore a 409 error from Taskcluster when calling reclaimTask for
a finished build. The build has already finished, and we can let
BBListener handle the finished event. In addition, if we don't ignore
the 409 error here, the generic taskcluster exception  handler will try
and cancel the buildbot build, which can take a long time.

Other minor changes:
- Submit timing information for how long it takes to refresh the list
of allowed builders

- Replace pytest-capturelog with pytest-catchlog to fix some pytest
warnings.